move more code out of Mimer.hs
